requirements.txt
torch
requests
transformers
Pillow
huggingface-hub
accelerate
peft
scikit-learn
pandas
sentencepiece
datasets
scipy

